requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.8.2
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
distlib==0.3.9
exceptiongroup==1.3.0
filelock==3.16.1
h11==0.16.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
kaitaistruct==0.10
numpy==1.24.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.0.3
platformdirs==4.3.6
py4j==0.10.9.7
pyasn1==0.6.1
pycparser==2.22
pyOpenSSL==25.0.0
pyparsing==3.1.4
PySocks==1.7.1
pyspark==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.32.3
selenium==4.27.1
selenium-wire==5.1.0
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.27.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2023.3
urllib3==2.2.3
virtualenv==20.31.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.23.0
